Enabled output logging on console and launch logs
@fabrizioromanelli I understood how the launch system works: by default, stdout is redirected to
a log file and only stderr is shown on screen. This is to address by default systems where
everything runs headlessly. However, this behaviour can be changed with the two options I added.
Now we should see everything ORB_SLAM2 does in the console and launch log file as well.
I/O buffering is disabled in orbslam2_ros2.cpp::main so everything happens asap.
If you have a couple of mins please try and check if this doesn't generate too many prints but a nice
console log.
Refs are:
- https://answers.ros.org/question/332829/no-stdout-logging-output-in-ros2-using-launch/
- https://github.com/ros2/launch/issues/188
- Python ROS 2 source code for ExecuteProcess class and launch.logging module