Added RealSense QoS profile with buffer size 1
@fabrizioromanelli this way the ORB_SLAM2 node listens for frames with a 1-sample buffer.
At least, considering the time it takes to process one, we should always have up-to-date information.
The buffer on the sender side (RealSense nodes) is still 5-samples large as per sensor_data QoS profile.
Problem is, if you look at the Intel code to try to change that as I just did, you'll want to tear
your eyes out. But considering delays and synchronization problems that might occur in our pipeline,
I believe it's better to leave it that way, so that no matter what happens we always have up to 5
frames to process. I'd like to hear your thoughts on this one.